+++ es/ESSubLong.cpp
@@ -2,11 +2,11 @@
 
 using namespace ES3;
 
-long ESSubLong::get(){
+uint32_t ESSubLong::get(){
 	return mLong;
 }
 
-void ESSubLong::set(long l){
+void ESSubLong::set(uint32_t l){
 	mRecordSize = 4;
 	mLong = l;
 }
@@ -14,6 +14,6 @@ void ESSubLong::set(long l){
 void ESSubLong::read(std::ifstream &ifs){
 
 	//read the size of the string
-	ifs.read ((char *)&mRecordSize, sizeof(long));
+	ifs.read ((char *)&mRecordSize, sizeof(uint32_t));
 	ifs.read((char*)&mLong, mRecordSize);
 }
