+++ es/ESSubRecord.cpp
@@ -7,7 +7,7 @@ void ESSubRecord::write(std::ofstream &ofs){}
 void ESSubRecord::read(std::ifstream &ifs){
 
 	//read the size of the string
-	ifs.read ((char *)&mRecordSize, sizeof(long));
+	ifs.read ((char *)&mRecordSize, sizeof(uint32_t));
 
 	char* tSubStr;
 	tSubStr = new char[mRecordSize + 2];
